+++ stateproofs/stateproofverification/merklearray/vectorCommitmentArray.go
@@ -1,4 +1,4 @@
-package stateproofverification
+package merklearray
 
 import (
 	"fmt"
